feat: 컴포넌트 구조 수정 및 뒷면 클릭시 앞면 보여주기 구현
