--- find-ompangi/src/components/Cards.jsx
@@ -1,140 +1,0 @@
-import { useEffect } from "react";
-import styled from "styled-components";
-
-const EASY = "EASY";
-const NORMAL = "NORMAL";
-const HARD = "HARD";
-
-const Cards = (props) => {
-  const OMPANGI_DATA = [
-    {
-      id: 1,
-      alt: "사랑에 빠진 옴팡이",
-      imgSrc:
-        "https://item.kakaocdn.net/do/f6223d3a6a7214e778b1dc54cd5160158b566dca82634c93f811198148a26065",
-    },
-    {
-      id: 2,
-      alt: "노트북 던지는 옴팡이",
-      imgSrc:
-        "https://item.kakaocdn.net/do/841caa712dcd88840271af09faf3586ff604e7b0e6900f9ac53a43965300eb9a",
-    },
-    {
-      id: 3,
-      alt: "드럼 치는 옴팡이",
-      imgSrc:
-        "https://item.kakaocdn.net/do/841caa712dcd88840271af09faf3586f41d1a2caccd0c566eab28b91e2e5d306",
-    },
-    {
-      id: 4,
-      alt: "자신감 넘치는 옴팡이",
-      imgSrc:
-        "https://item.kakaocdn.net/do/4d1df026331431133006180113d5c4d1effd194bae87d73dd00522794070855d",
-    },
-    {
-      id: 5,
-      alt: "이불 덮고 우는 옴팡이",
-      imgSrc:
-        "https://item.kakaocdn.net/do/480b76dbaaedeeacf3904a31edf195269f17e489affba0627eb1eb39695f93dd",
-    },
-    {
-      id: 6,
-      alt: "팬더 옴팡이",
-      imgSrc:
-        "https://item.kakaocdn.net/do/511631850c4926152b31ed2db55b4d80ac8e738cb631e72fdb9a96b36413984e",
-    },
-    {
-      id: 7,
-      alt: "서러운 옴팡이",
-      imgSrc:
-        "https://item.kakaocdn.net/do/511631850c4926152b31ed2db55b4d80339e41ce89b663315d96faecd7cfd11b",
-    },
-    {
-      id: 8,
-      alt: "화난 옴팡이",
-      imgSrc:
-        "https://item.kakaocdn.net/do/511631850c4926152b31ed2db55b4d80ce9463e040a07a9462a54df43e1d73f1",
-    },
-    {
-      id: 9,
-      alt: "행운을 비는 옴팡이",
-      imgSrc:
-        "https://item.kakaocdn.net/do/fec24d1252d804d038b5a3a6304bee83113e2bd2b7407c8202a97d2241a96625",
-    },
-  ];
-
-  const { level } = props;
-
-  // 배열 셔플 함수
-  const shuffle = (array) => {
-    array.sort(() => Math.random() - 0.5);
-  };
-
-  // 원본 배열 셔플
-  shuffle(OMPANGI_DATA);
-  const slicedData = OMPANGI_DATA.slice(
-    0,
-    level === EASY ? 5 : level === NORMAL ? 7 : 9
-  );
-  const renderData = [...slicedData, ...slicedData];
-  shuffle(renderData);
-
-  // 카드 렌더링 --> key 값 고민
-  const cardList = renderData.map(({ id, alt, imgSrc }, idx) => (
-    <St.CardFront key={`${id}_${idx}`} className={id}>
-      <img src={imgSrc} alt={alt} />
-    </St.CardFront>
-  ));
-
-  return <St.CardContainer>{cardList}</St.CardContainer>;
-};
-
-export default Cards;
-
-const St = {
-  CardContainer: styled.section`
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    row-gap: 2rem;
-
-    width: 73%;
-    margin-top: 1rem;
-  `,
-
-  CardFront: styled.article`
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    width: 19rem;
-    height: 25rem;
-
-    border-radius: 1rem;
-
-    background-color: white;
-
-    cursor: pointer;
-
-    & > img {
-      width: 15rem;
-    }
-  `,
-
-  CardBack: styled.article`
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    width: 19rem;
-    height: 25rem;
-
-    font-size: 2.7rem;
-
-    border-radius: 1rem;
-
-    background-color: ${({ theme }) => theme.colors.darkPink};
-
-    cursor: pointer;
-  `,
-};
